backend/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
boto3==1.26.21
cryptography==38.0.4
ipython==8.7.0
python-decouple==3.6
web3==5.31.2
gunicorn==20.1.0
gevent==22.10.2
psycopg2==2.9.1
django-cors-headers==3.13.0
celery==5.2.7
redis==4.4.0
